refactor: remove unused dao
