+++ local_datasource/src/main/java/com/baghdad/local_datasource/database/dao/MovieDao.kt
@@ -1,9 +1,9 @@
-package com.baghdad.local_datasource.roomDB.dao
+package com.baghdad.local_datasource.database.dao
 
 import androidx.room.Dao
 import androidx.room.Query
 import androidx.room.Upsert
-import com.baghdad.local_datasource.roomDB.entity.Movie
+import com.baghdad.local_datasource.database.entity.Movie
 import kotlinx.coroutines.flow.Flow
 
 @Dao
