+++ local_datasource/src/main/java/com/baghdad/local_datasource/database/dao/RecentSearchDao.kt
@@ -1,9 +1,9 @@
-package com.baghdad.local_datasource.roomDB.dao
+package com.baghdad.local_datasource.database.dao
 
 import androidx.room.Dao
 import androidx.room.Query
 import androidx.room.Upsert
-import com.baghdad.local_datasource.roomDB.entity.RecentSearch
+import com.baghdad.local_datasource.database.entity.RecentSearch
 import kotlinx.coroutines.flow.Flow
 
 @Dao
@@ -12,25 +12,25 @@ interface RecentSearchDao {
     @Upsert
     fun addRecentSearch(recentSearch: RecentSearch)
 
-    @Query("SELECT * FROM RecentSearch")
+    @Query("SELECT * FROM RecentlySearch")
     fun getAllRecentSearch(): Flow<List<RecentSearch>>
 
-    @Query("SELECT * FROM RecentSearch WHERE id = :id")
+    @Query("SELECT * FROM RecentlySearch WHERE id = :id")
     suspend fun getRecentSearchById(id: Long): List<RecentSearch>
 
-    @Query("SELECT * FROM RecentSearch ORDER BY searchedAt DESC")
+    @Query("SELECT * FROM RecentlySearch ORDER BY searchedAt DESC")
     fun getLastTenRecentSearchItems(): Flow<List<RecentSearch>>
 
 
-    @Query("DELETE FROM RecentSearch")
+    @Query("DELETE FROM RecentlySearch")
     suspend fun clearAllRecentSearch()
 
-    @Query("SELECT COUNT(*) FROM RecentSearch")
+    @Query("SELECT COUNT(*) FROM RecentlySearch")
     suspend fun getCount(): Int
 
-    @Query("SELECT EXISTS(SELECT 1 FROM RecentSearch WHERE id = :id)")
+    @Query("SELECT EXISTS(SELECT 1 FROM RecentlySearch WHERE id = :id)")
     suspend fun exists(id: Long): Boolean
 
-    @Query("DELETE FROM RecentSearch WHERE id = :id")
+    @Query("DELETE FROM RecentlySearch WHERE id = :id")
     suspend fun deleteRecentSearchById(id: Long)
 }
